requirements.txt
streamlit>=1.28
opencv-python-headless>=4.8
face-recognition>=1.3
numpy>=1.24
pandas>=2.0
Pillow>=10.0
